# Conflicts:
#	proyecto2/core/src/com/mygdx/game/PlayScreen.java

+++ proyecto2/core/src/com/mygdx/game/sprites/Jugador.java
@@ -34,7 +34,7 @@ public class Jugador implements Serializable {
         this.nombre = nombre;
     }
 
-    public int getScore() {
+    public float getScore() {
         return score;
     }
